requirements.txt
tornado==4.0.2
websocket-client
pyyaml
nose
MySQL-python
